# src/Kentico.Xperience.K13Ecommerce/StoreApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kentico Xperience 13 Store API",
    "version": "v1"
  },
  "paths": {
...
    "/api/store/cart/content": {
      "get": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "GetCurrentCartContent",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KShoppingCartContent"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/details": {
      "get": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "GetCurrentCartDetails",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KShoppingCartDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/customer": {
      "get": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "GetCurrentCustomer",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KCustomerShoppingCartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/add-item": {
      "post": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "AddItemToCart",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skuId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "schema": {
              "maximum": 32767,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KShoppingCartItemShoppingCartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/update-item": {
      "put": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "UpdateItemQuantity",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 32767,
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/remove-item": {
      "delete": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "RemoveItemFromCart",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/add-coupon-code": {
      "post": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "couponCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanShoppingCartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/remove-coupon-code": {
      "delete": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "couponCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/store/cart/set-payment-option": {
      "put": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "paymentOptionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/store/cart/set-customer": {
      "put": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KCustomer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KCustomer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KCustomer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/set-billing-address": {
      "put": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KAddress"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KAddress"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KAddress"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingCartBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/store/cart/summary": {
      "get": {
        "tags": [
          "ShoppingCart"
        ],
        "operationId": "GetCurrentCartSummary",
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KShoppingCartSummary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/store/cart/create-order": {
      "post": {
        "tags": [
          "ShoppingCart"
        ],
        "parameters": [
          {
            "name": "ShoppingCartGUID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "note": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KOrder"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KCustomerShoppingCartResponse": {
        "type": "object",
        "properties": {
          "shoppingCartGuid": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "$ref": "#/components/schemas/KCustomer"
          }
        },
        "additionalProperties": false
      },
...
      "KShoppingCartContent": {
        "type": "object",
        "properties": {
          "shoppingCartGuid": {
            "type": "string",
            "format": "uuid"
          },
          "grandTotal": {
            "type": "number",
            "format": "decimal"
          },
          "totalPrice": {
            "type": "number",
            "format": "decimal"
          },
          "totalItemsPrice": {
            "type": "number",
            "format": "decimal"
          },
          "otherPayments": {
            "type": "number",
            "format": "decimal"
          },
          "orderDiscount": {
            "type": "number",
            "format": "decimal"
          },
          "taxSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KSummaryItem"
            },
            "nullable": true
          },
          "orderDiscountSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KSummaryItem"
            },
            "nullable": true
          },
          "itemsDiscount": {
            "type": "number",
            "format": "decimal"
          },
          "totalTax": {
            "type": "number",
            "format": "decimal"
          },
          "totalShipping": {
            "type": "number",
            "format": "decimal"
          },
          "remainingAmountForFreeShipping": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "$ref": "#/components/schemas/KCurrency"
          },
          "couponCodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ICouponCode"
            },
            "nullable": true
          },
          "cartProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KShoppingCartItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KShoppingCartDetails": {
        "required": [
          "billingAddress",
          "customer",
          "shoppingCartGuid"
        ],
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/KCustomer"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/KAddress"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/KAddress"
          },
          "shippingOptionId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentOptionId": {
            "type": "integer",
            "format": "int32"
          },
          "shoppingCartGuid": {
            "type": "string",
            "format": "uuid"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "isShippingNeeded": {
            "type": "boolean"
          },
          "availableShippingOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KShippingOption"
            },
            "nullable": true
          },
          "availablePaymentOptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KPaymentOption"
            },
            "nullable": true
          },
          "customData": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KShoppingCartItemShoppingCartResponse": {
        "type": "object",
        "properties": {
          "shoppingCartGuid": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "$ref": "#/components/schemas/KShoppingCartItem"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ShoppingCartBaseResponse": {
        "type": "object",
        "properties": {
          "shoppingCartGuid": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}